+++ .travis.yml
@@ -1,0 +1,38 @@
+language: dart
+dart:
+  - stable
+os:
+  - linux
+sudo: false
+addons:
+  apt:
+    sources:
+      - ubuntu-toolchain-r-test # you need this source to get the right version of libstdc++6
+    packages:
+      - libstdc++6
+      - fonts-droid
+install:
+  - echo 'Avoid default Travis CI install step' # this is to avoid an error with pub in Travis
+before_script:
+  - cd ..
+  - git clone https://github.com/flutter/flutter.git
+  - export PATH=`pwd`/flutter/bin:`pwd`/flutter/bin/cache/dart-sdk/bin:$PATH
+  - flutter doctor
+script:
+  - cd $TRAVIS_BUILD_DIR
+  - flutter packages get
+  - flutter analyze --no-pub --no-current-package lib
+  - flutter test
+  - flutter packages pub publish --dry-run
+before_deploy:
+  - chmod +x ./.travis/publish.sh # giving execution permissions to this file to avoid error 127.
+  - mv ./.travis ../ # moving this out the publication folder as we don't want to publish it.
+deploy:
+  provider: script
+  skip_cleanup: true
+  script: "../.travis/publish.sh"
+  on:
+    tags: true
+cache:
+  directories:
+    - $HOME/.pub-cache
