+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.1]
+
+    - add video descriptions
+
 ## [1.0.0]
 
     - initial release
